ver 1.5.1
Now the datalog strategy is changed: before it ends after a certain amount of fixed time; now it ends after detecting a certain number of taps. This means that the time can vary.
Also, the APP will send a specific signal to the device to let the device stop datalogging too.

(Probably) TODO: The datalog stops immediately when detecting the last tap. Maybe a delay is needed.